// src/app/shared/interfaces/motivo.model.ts
export interface MotivoMV {
  id?: number;
  descricaoMotivo?: string;
  descricaoTipoAtendimento?: string;
  tipoAtendimentoSolicitacao?: string;
  atendimentoTipoId?: number;
  atendimentoSolicitacaoTipoId?: number;
  descricaoMotivoSolicitacao?: string;
  categoriasFornecedorRaio?: string;
  inseridoEm?: string;
  inseridoPor?: number;
  modificadoEm?: string;
  modificadoPor?: number;
  nomeModificadoPor?: string;
  nomeInseridoPor?: string;
  nomeAtendimentoTipo?: string;
}